Paraxial Analysis tools Fixes #242
wrote test programes for virtual point and projection. Moved beam energy code into lenslet assembly.
